Moved extraction into own file, and use string processing to locate valid rel attributes instead of relying on goquery selectors.
